+++ NEWS.md
@@ -7,7 +7,9 @@
 ENHANCEMENT :
 
 * `importZipStudyWeb()` zip study with new package `zip`instead of `utils::zip` (more faster)
+* `createLink()`/ `editLink()` check the number of rows of dataLink and tsLink if provided
 * `updateOptimizationSettings()` add missing properties and update possible values by property
+
 
 BUGFIXES :